requirements.txt
httpx
tqdm
termcolor
python-nmap
dnspython
PyYAML
packaging
requests